# Conflicts:
#	.clabot
#	Dockerfile
#	docker-compose.async.yml
#	docker-compose.yml
#	docs/basic/settings/index.rst
#	docs/intermediate/monitoring/geonode_monitoring/index.rst
#	geonode/__init__.py
#	geonode/api/resourcebase_api.py
#	geonode/base/models.py
#	geonode/base/templates/base/thumbnail_upload.html
#	geonode/documents/templates/layouts/doc_panels.html
#	geonode/geoserver/helpers.py
#	geonode/groups/tests.py
#	geonode/layers/templates/layouts/panels.html
#	geonode/layers/views.py
#	geonode/local_settings.py.geoserver.sample
#	geonode/maps/templates/layouts/map_panels.html
#	geonode/maps/views.py
#	geonode/monitoring/tests/integration.py
#	geonode/people/signals.py
#	geonode/proxy/templatetags/proxy_lib_tags.py
#	geonode/proxy/views.py
#	geonode/settings.py
#	geonode/static/lib/css/assets.min.css
#	geonode/static/lib/css/bootstrap-multiselect.css
#	geonode/static/lib/css/bootstrap-select.css
#	geonode/static/lib/css/bootstrap-table.css
#	geonode/static/lib/css/jquery-ui.theme.css
#	geonode/static/lib/js/affix.js
#	geonode/static/lib/js/alert.js
#	geonode/static/lib/js/angular.js
#	geonode/static/lib/js/angular.min.js
#	geonode/static/lib/js/assets.min.js
#	geonode/static/lib/js/bootstrap-multiselect.js
#	geonode/static/lib/js/bootstrap-select.js
#	geonode/static/lib/js/bootstrap-table.js
#	geonode/static/lib/js/bootstrap.js
#	geonode/static/lib/js/bootstrap.min.js
#	geonode/static/lib/js/button.js
#	geonode/static/lib/js/carousel.js
#	geonode/static/lib/js/collapse.js
#	geonode/static/lib/js/dropdown.js
#	geonode/static/lib/js/handlebars.js
#	geonode/static/lib/js/handlebars.min.js
#	geonode/static/lib/js/leaflet-plugins.min.js
#	geonode/static/lib/js/leaflet.js
#	geonode/static/lib/js/modal.js
#	geonode/static/lib/js/moment-timezone-with-data.js
#	geonode/static/lib/js/popover.js
#	geonode/static/lib/js/scrollspy.js
#	geonode/static/lib/js/tab.js
#	geonode/static/lib/js/tooltip.js
#	geonode/static/lib/js/transition.js
#	geonode/static/package.json
#	geonode/templates/500.html
#	geonode/templates/base.html
#	geonode/templates/metadata_base.html
#	geonode/templates/search/search_scripts.html
#	geonode/utils.py
#	requirements.txt
#	scripts/spcgeonode/django/Dockerfile

+++ geonode/security/tests.py
@@ -204,15 +204,27 @@ class SecurityViewsTests(ResourceTestCaseMixin, GeoNodeBaseTestSupport):
             'uuid': test_layer.uuid
         }
         resp = self.client.post(reverse('attributes_sats_refresh'), data)
-        self.assertHttpOK(resp)
-        self.assertEquals(layer_attributes.count(), test_layer.attributes.count())
+        if resp.status_code == 200:
+            self.assertHttpOK(resp)
+            self.assertEquals(layer_attributes.count(), test_layer.attributes.count())
 
-        from geonode.geoserver.helpers import set_attributes_from_geoserver
-        test_layer.attribute_set.all().delete()
-        test_layer.save()
+            from geonode.geoserver.helpers import set_attributes_from_geoserver
+            test_layer.attribute_set.all().delete()
+            test_layer.save()
 
-        set_attributes_from_geoserver(test_layer, overwrite=True)
-        self.assertEquals(layer_attributes.count(), test_layer.attributes.count())
+            set_attributes_from_geoserver(test_layer, overwrite=True)
+            self.assertEquals(layer_attributes.count(), test_layer.attributes.count())
+
+            # Remove permissions to anonymous users and try to refresh attributes again
+            test_layer.set_permissions({'users': {'AnonymousUser': []}})
+            test_layer.attribute_set.all().delete()
+            test_layer.save()
+
+            set_attributes_from_geoserver(test_layer, overwrite=True)
+            self.assertEquals(layer_attributes.count(), test_layer.attributes.count())
+        else:
+            # If GeoServer is unreachable, this view now returns a 302 error
+            self.assertEqual(resp.status_code, 302)
 
         # Remove permissions to anonymous users and try to refresh attributes again
         test_layer.set_permissions({'users': {'AnonymousUser': []}})
